Fix abi_type_strategies.sized_list_strats
The values of `type_str` and `type_strat` are not captured in the
closures of each lambda.  Therefore, every resulting lambda references
the values in the last tuple in `all_basic_raw_strats` which happen to
be for address sampling.

Also, the shared value key should probably be more unique.